Modularize routes.ts wave 5: me + knowledge + orchestrate-misc
server/routes-modules/me.ts (108)
  /api/me, /api/settings/personalization (GET + PUT),
  /api/me/avatar (multipart upload + session sync + DB upsert).

server/routes-modules/knowledge.ts (267)
  All 12 /api/knowledge/* endpoints: context build, search,
  core-memory KV, project-memory CRUD, personal-base profile,
  scratchpad CRUD. Carries its own ensureSessionUserInDatabase
  helper so the FK-bearing memory inserts don't trip the users
  table.

server/routes-modules/orchestrate-and-misc.ts (192)
  /api/orchestrate, /api/orchestrate/status,
  /api/voice/transcribe, /api/admin/knowledge/overview,
  /api/admin/system-test, /api/chat. Packed together because
  each handler is small and they share dependencies. Takes the
  isDatabaseHealthy getter (same pattern as diagnostics).

routes.ts: 1,089 → 670 (-419 lines, -38%).
Combined waves 1-5: 2,101 → 670 (-68%).

Only the conversation routes remain inline (10 endpoints,
including the big /api/conversations/:id/messages SSE handler).
Final wave extracts those.
